state: use hashmap with UUID key for clients
And change connmgr to use UUID key also.

Depending on the Vec index doesn't work, as once a client is deleted,
the Vec index for the clients after the deleted client will be one
lower, resulting in a mismatch between the client ID in connmgr.

Let's use a UUID for now.

Once we implement database support, we can assign an ID to a Willow when
it first connects, save it in the database, and reuse that ID. This way
we can also replace the client in the state, should it still be there.

As the UI expects an array of clients from GET /api/clients, we need to
modify the get_api_client function to return a Vec.

Signed-off-by: Stijn Tintel <stijn@linux-ipv6.be>