willow: add custom deserializers for config
The Python version stores string, int and bool in the same VARCHAR
column in the database, and converts it back to the original type by
mapping the database query result to a Pydantic model.

Unfortunately we cannot use sqlx query_as to automagically do the same.
We therefore need a few custom deserializers.

Signed-off-by: Stijn Tintel <stijn@linux-ipv6.be>