websocket: check User-Agent in get_ws
If we're going to deny clients without User-Agent header, it's more
efficient to do it as early as possible. We can also return a standard
HTTP status code, and we use less memory in handle_ws, which is spawned
for every client connected to the WebSocket.

As we're now inserting adding the client to the state in get_ws, we need
to remove it in case the WebSocket upgrade fails. This is a bit tricky,
as Axum on_failed_upgrade does not take an async function.

Signed-off-by: Stijn Tintel <stijn@linux-ipv6.be>